Minor cleanup in relation to container copying and serialization and make a bit of progress towards having chemicals be ints so that we can avoid having to clamp values

+++ src/api/java/mekanism/api/chemical/ChemicalResource.java
@@ -136,7 +136,7 @@ public final class ChemicalResource implements RegisteredResource<Chemical>, IHa
      *
      * @throws IllegalArgumentException when amount is negative.
      */
-    public ChemicalStack toStack(long amount) {
+    public ChemicalStack toStack(int amount) {
         MekanismPreconditions.checkNonNegative(amount);
         if (amount == 0) {
             return ChemicalStack.EMPTY;
